[fix:IS-2] rename from styles to styled
