+++ package.json
@@ -71,7 +71,7 @@
     "workbox-webpack-plugin": "6.4.1"
   },
   "scripts": {
-    "start": "react-app-rewired start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
     "eject": "react-scripts eject"
