+++ src/components/AuthCardInfo/styled.ts
